+++ .github/workflows/deploy.yml
@@ -27,7 +27,8 @@ jobs:
           sudo apt-get install -y libimage-exiftool-perl
           python -m pip install --upgrade pip
           pip install pytest
-          pip install -r requirements-dev.txt
+          pip install -r requirements.lock
+          pip install -r requirements-dev.lock
 
       - name: Run tests
         run: pytest
