+++ .github/workflows/tests.yml
@@ -29,7 +29,8 @@ jobs:
         run: |
           sudo apt-get install -y libimage-exiftool-perl
           python -m pip install --upgrade pip
-          pip install -r requirements-dev.txt
+          pip install -r requirements.lock
+          pip install -r requirements-dev.lock
 
       - name: Run tests
         run: pytest
