+++ .gitignore
@@ -15,7 +15,7 @@ __pycache__/
 *.py[cod]
 *$py.class
 .pytest_cache
-
+.ruff_cache
 # C extensions
 *.so
