+++ phockup/internal/exif.py
