+++ pyproject.toml
@@ -1,0 +1,36 @@
+[project]
+name = "phockup"
+version = "1.13.0"
+description = "Media sorting tool to organize photos and videos from your camera in folders by year, month and day"
+authors = [
+    { name = "Ivan Dokov", email = "ivan@jetspark.io" }
+]
+dependencies = [
+    "tqdm"
+]
+readme = "readme.md"
+requires-python = ">= 3.8"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.rye]
+universal = true
+managed = true
+dev-dependencies = [
+    "pytest>=8.3.3",
+    "pytest-mock",
+    "pytest-cov",
+    "pytest-socket",
+    "pre-commit",
+]
+
+[tool.hatch.metadata]
+allow-direct-references = true
+
+[project.scripts]
+phockup = 'phockup:app'
+
+[tool.ruff.lint]
+ignore = ["E203", "E501", "W503", "E126"]
