+++ requirements-dev.lock
@@ -1,0 +1,54 @@
+# generated by rye
+# use `rye lock` or `rye sync` to update this lockfile
+#
+# last locked with the following flags:
+#   pre: false
+#   features: []
+#   all-features: false
+#   with-sources: false
+#   generate-hashes: false
+#   universal: true
+
+-e file:.
+cfgv==3.4.0
+    # via pre-commit
+colorama==0.4.6 ; sys_platform == 'win32' or platform_system == 'Windows'
+    # via pytest
+    # via tqdm
+coverage==7.6.1
+    # via pytest-cov
+distlib==0.3.8
+    # via virtualenv
+exceptiongroup==1.2.2 ; python_full_version < '3.11'
+    # via pytest
+filelock==3.16.0
+    # via virtualenv
+identify==2.6.0
+    # via pre-commit
+iniconfig==2.0.0
+    # via pytest
+nodeenv==1.9.1
+    # via pre-commit
+packaging==24.1
+    # via pytest
+platformdirs==4.3.2
+    # via virtualenv
+pluggy==1.5.0
+    # via pytest
+pre-commit==3.5.0
+pytest==8.3.3
+    # via pytest-cov
+    # via pytest-mock
+    # via pytest-socket
+pytest-cov==5.0.0
+pytest-mock==3.14.0
+pytest-socket==0.7.0
+pyyaml==6.0.2
+    # via pre-commit
+tomli==2.0.1 ; python_full_version <= '3.11'
+    # via coverage
+    # via pytest
+tqdm==4.66.5
+    # via phockup
+virtualenv==20.26.4
+    # via pre-commit
