--- requirements-dev.txt
@@ -1,6 +1,0 @@
--r requirements.txt
-pytest
-pytest-mock
-pytest-cov
-pytest-socket
-pre-commit
