+++ requirements.lock
@@ -1,0 +1,16 @@
+# generated by rye
+# use `rye lock` or `rye sync` to update this lockfile
+#
+# last locked with the following flags:
+#   pre: false
+#   features: []
+#   all-features: false
+#   with-sources: false
+#   generate-hashes: false
+#   universal: true
+
+-e file:.
+colorama==0.4.6 ; platform_system == 'Windows'
+    # via tqdm
+tqdm==4.66.5
+    # via phockup
