--- requirements.txt
@@ -1,1 +1,0 @@
-tqdm==4.66.1
