+++ tests/test_date.py
@@ -3,7 +3,7 @@ import os
 import re
 from datetime import datetime
 
-from src.date import Date
+from phockup.internal.date import Date
 
 os.chdir(os.path.dirname(__file__))
