--- tox.ini
@@ -1,3 +1,0 @@
-[flake8]
-max-line-length = 88
-ignore = E203 E501 W503 E126
